% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkmode.R
\name{with_darkmode}
\alias{with_darkmode}
\title{Title}
\usage{
with_darkmode(bottom = "32px", right = "32px", left = "unset",
  time = "0.3s", mixColor = "#fff", backgroundColor = "#fff",
  buttonColorDark = "#100f2c", buttonColorLight = "#fff",
  saveInCookies = TRUE, label = "", autoMatchOsTheme = TRUE,
  online = TRUE)
}
\arguments{
\item{bottom, right, left, time, mixColor, backgroundColor, buttonColorDark, buttonColorLight, saveInCookies, label, autoMatchOsTheme}{Parameters to pass to Darkmode.js}

\item{online}{Use CDN or local version}
}
\description{
Title
}
